style: clippy --workspace --all-targets mechanical fixes
Apply all autofix-style and Category D lints across the workspace:
- field_reassign_with_default → struct-literal init with ..Default::default()
- doc_lazy_continuation / doc_overindented_list_items → fix comment indentation
- while_let_loop → rewrite loop+match as while-let in scheduler
- never_loop → rewrite for+break as if-let in drm_direct
- unnecessary_cast/unwrap/parens/to_owned/clone_on_double_ref → trivial rewrites
- erasing_op → replace 0*cols+1 with literal 1 in gpu_pipeline test
- assertions_on_constants → const { assert!() } or delete sanity comment
- len_without_is_empty → add is_empty() to FragmentCoverageMap
- needless_update → remove redundant ..Default::default() on single-field structs
- needless_range_loop in pal_rle_tests/gpu_pipeline/palette_churn → enumerate()
- same_item_push → extend(repeat_n)
- double_parens / unusual_byte_groupings → cosmetic fixes
- filter_next → rfind in e2e test
- dead setup_e2e* helpers in e2e.rs removed (never called)
- too_many_arguments in gpu_pipeline integration test → ChangedMemfdParams struct
- cdf53.rs: #![allow(needless_range_loop)] + doc comment fixes

+++ ghostframe-test-pattern/src/mode_switch.rs
@@ -54,7 +54,7 @@ pub fn run<C: Connection>(
         let stripe1_y = 0i16;
         let stripe2_y = stripe_h as i16;
         let stripe3_y = (stripe_h * 2) as i16;
-        let stripe3_h = (height - stripe_h * 2) as u32;
+        let stripe3_h = height - stripe_h * 2;
 
         // Stripe 1 — solid.
         conn.change_gc(gc, &ChangeGCAux::default().foreground(STATIC_BG_PIXEL))?;
@@ -101,9 +101,9 @@ pub fn run<C: Connection>(
         // the classifier just needs > 16 colors).
         for row in 0..stripe3_h {
             let py = stripe3_y as u32 + row;
-            let r = (py.wrapping_mul(3) & 0xFF) as u32;
-            let g = (py.wrapping_mul(5) & 0xFF) as u32;
-            let b = (py.wrapping_mul(7) & 0xFF) as u32;
+            let r = py.wrapping_mul(3) & 0xFF;
+            let g = py.wrapping_mul(5) & 0xFF;
+            let b = py.wrapping_mul(7) & 0xFF;
             let color = (r << 16) | (g << 8) | b;
             conn.change_gc(gc, &ChangeGCAux::default().foreground(color))?;
             conn.poly_fill_rectangle(
